educational-multiplex-gnn/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
torch-geometric>=2.3.0
torch-scatter>=2.1.2
torch-sparse>=0.6.18
torch-cluster>=1.6.3
torch-spline-conv>=1.2.2
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.0
pyyaml>=6.0
tqdm>=4.65.0
shap>=0.42.0
umap-learn>=0.5.0
xgboost>=2.0.0
ray[tune]>=2.5.0
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
